use constants for number of bits, return correct type

+++ k4FWCore/include/k4FWCore/IUniqueIDGenSvc.h
@@ -11,9 +11,9 @@
 
 struct GAUDI_API IUniqueIDGenSvc : extend_interfaces<IInterface> {
   DeclareInterfaceID(IUniqueIDGenSvc, 2, 0);
-  virtual uint64_t getUniqueID(uint32_t evt_num) const = 0;
-  virtual uint64_t getUniqueID(uint32_t evt_num, uint32_t run_num) const = 0;
-  virtual uint64_t getUniqueID(uint32_t evt_num, uint32_t run_num, std::string name) const = 0;
+  virtual size_t getUniqueID(uint32_t evt_num) const = 0;
+  virtual size_t getUniqueID(uint32_t evt_num, uint32_t run_num) const = 0;
+  virtual size_t getUniqueID(uint32_t evt_num, uint32_t run_num, std::string name) const = 0;
 };
 
 #endif
